+++ src/android/Options.java
@@ -95,9 +95,9 @@ public final class Options {
             if (isVersionOlder(metaVersion, "1.1.1")) convertPropertiesForVersion111(options);
             if (isVersionOlder(metaVersion, "1.1.8")) convertPropertiesForVersion118(options);
 
-            // Update meta.version always if it is older
-            if (isVersionOlder(metaVersion, "1.2.1-dev")) {
-                meta.put("version", "1.2.1-dev");
+            // Update meta.version to current plugin version
+            if (!metaVersion.equals("1.2.1")) {
+                meta.put("version", "1.2.1");
                 options.put("meta", meta);
             }
         } catch (JSONException exception) {