+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Bugfix: Handle crash when updating from plugin version `0.9-beta.3` and notifications were already scheduled with the old plugin version.
   - This was due to not handling correctly the `-beta` in the version string `0.9-beta.3` when trying to convert the string to an int. Now the pre-release identifiers `-dev` and `-beta` will be removed, before the version string will be converted to an int.
   - Fixes [#2087](https://github.com/katzer/cordova-plugin-local-notifications/issues/2087)
+- Update `meta.version` of scheduled notifications from older plugin versions to the current plugin version. So if notifications were scheduled by version `1.2.0`, it will be updated to `1.2.1`, when the app is updated.
 
 ## Version 1.2.0 (02.07.2025)